+++ Sources/StreamVideoSwiftUI/CallView/CallControls/Stateless/StatelessToggleCameraIconView.swift
@@ -15,7 +15,7 @@ public struct StatelessToggleCameraIconView: View {
     @Injected(\.images) private var images
 
     /// The associated call for the toggle camera icon.
-    public var call: Call?
+    public weak var call: Call?
 
     /// The size of the toggle camera icon.
     public var size: CGFloat