+++ StreamVideoTests/Utils/AudioSession/StreamAudioSession_Tests.swift
@@ -239,6 +239,76 @@ final class StreamAudioSession_Tests: XCTestCase, @unchecked Sendable {
         XCTAssertNil(StreamAudioSession.currentValue)
     }
 
+    // MARK: - callKitActivated
+
+    func test_callKitActivated_configurationWasCalledOnPolicy() async throws {
+        let mockPolicy = MockAudioSessionPolicy()
+        try await subject.didUpdatePolicy(mockPolicy)
+        let audioSession = MockAVAudioSession()
+
+        try subject.callKitActivated(audioSession)
+
+        // The expected value is 2 as the audioSession will call it once
+        // when we first update the policy.
+        XCTAssertEqual(mockPolicy.timesCalled(.configuration), 2)
+    }
+
+    func test_callKitActivated_providedAudioSessionSetCategoryWasCalledCorrectly() async throws {
+        let mockPolicy = MockAudioSessionPolicy()
+        mockPolicy.stub(
+            for: .configuration,
+            with: AudioSessionConfiguration(
+                category: .playAndRecord,
+                mode: .voiceChat,
+                options: .mixWithOthers,
+                overrideOutputAudioPort: .speaker
+            )
+        )
+        try await subject.didUpdatePolicy(mockPolicy)
+        let audioSession = MockAVAudioSession()
+
+        try subject.callKitActivated(audioSession)
+
+        let request = try XCTUnwrap(
+            audioSession.recordedInputPayload(
+                (
+                    AVAudioSession.Category,
+                    AVAudioSession.Mode,
+                    AVAudioSession.CategoryOptions
+                ).self,
+                for: .setCategory
+            )?.first
+        )
+        XCTAssertEqual(request.0, .playAndRecord)
+        XCTAssertEqual(request.1, .voiceChat)
+        XCTAssertTrue(request.2.contains(.mixWithOthers))
+    }
+
+    func test_callKitActivated_providedAudioSessionSetOverridePortWasCalledCorrectly() async throws {
+        let mockPolicy = MockAudioSessionPolicy()
+        mockPolicy.stub(
+            for: .configuration,
+            with: AudioSessionConfiguration(
+                category: .playAndRecord,
+                mode: .voiceChat,
+                options: .mixWithOthers,
+                overrideOutputAudioPort: .speaker
+            )
+        )
+        try await subject.didUpdatePolicy(mockPolicy)
+        let audioSession = MockAVAudioSession()
+
+        try subject.callKitActivated(audioSession)
+
+        let request = try XCTUnwrap(
+            audioSession.recordedInputPayload(
+                AVAudioSession.PortOverride.self,
+                for: .setOverrideOutputAudioPort
+            )?.first
+        )
+        XCTAssertEqual(request, .speaker)
+    }
+
     // MARK: - Private Helpers
 
     private func assertConfigurationWasCalledOnPolicy(